requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-timeout==1.3.4
requests==2.23.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
